Move file helpers to another location
