+++ src/pages/registration/files_and_license_tab/FilesTableRow.tsx
@@ -49,11 +49,11 @@ import { licenses, LicenseUri } from '../../../types/license.types';
 import { SpecificFileFieldNames } from '../../../types/publicationFieldNames';
 import { Registration } from '../../../types/registration.types';
 import { dataTestId } from '../../../utils/dataTestIds';
+import { hasFileAccessRight } from '../../../utils/fileHelpers';
 import { equalUris } from '../../../utils/general-helpers';
 import { isOpenFile, isPendingOpenFile, userIsValidImporter } from '../../../utils/registration-helpers';
 import { IdentifierParams } from '../../../utils/urlPaths';
 import { DeleteIconButton } from '../../messages/components/DeleteIconButton';
-import { hasFileAccessRight } from '../helpers/fileHelpers';
 import { DownloadFileButton } from './DownloadFileButton';
 
 const StyledFileTypeMenuItemContent = styled('div')({
